requirements.txt
streamlit==1.28.1
pandas==2.0.3
numpy==1.24.3
python-pptx==0.6.21
openpyxl==3.1.2
